Fixes crash Neutron OVS Agent to set tags for ports on Windows
The Neutron OVS agent on Windows does not set the vlan/vxlan
tags for the VM ports when using the native implementation.
(using ovsdb_interface = native and of_interface = native)

This happens because: the agent will try to fetch a row
by record. It will check to see if it is a UUID, if the condition
is true it will try to fetch the column directly.

On Windows the name of the ports is a UUID (for Windows
there is no 16 chars length limitation as for Linux).
Since the UUID check will pass, it will try to fetch the
column directly instead of using the lookup table on
Windows and will raise RowNotFound exception.

This patch addresses this issue by removing the exception
RowNotFound in the Windows case; the rule will be fetched
using the lookup table.

Change-Id: Ie0a980eaf2a6956072c2536274252764aadd227d
Closes-Bug: #1670451